CMS-850: Move helper file to the root of backend
