+++ backend/utils/firstComeFirstServedHelper.js
@@ -1,18 +1,25 @@
-import { DateRange, DateType } from "../../models/index.js";
+import { DateRange, DateType } from "../models/index.js";
 
 // FCFS dates calculation helper function
 
-// Example 1: Operation: 2025-03-21 to 2025-10-13
-//            Reservation: 2025-03-28 to 2025-10-12
-//            First come: 2025-03-21 to 2025-03-27
-// Example 2: Operation: 2025-04-01 to 2025-10-31
-//            Reservation: 2025-05-10 to 2025-10-12
-//            First come: 2025-04-01 to 2025-05-09
-//            First come: 2025-10-13 to 2025-10-30
+// Example 1
+// Reservation dates are in the range of Operation dates
 
-/**
- * @param {Object} season
- */
+// Operation: 2025-03-21 to 2025-10-13
+// Reservation: 2025-03-28 to 2025-10-12
+// First come: 2025-03-21 to 2025-03-27
+// => From Operation start date to Reservation start date minus 1
+
+// Example 2
+// Operation dates start earlier than Reservation dates
+// Operation dates end later than Reservation dates
+
+// Operation: 2025-04-01 to 2025-10-31
+// Reservation: 2025-05-10 to 2025-10-12
+// First come 1: 2025-04-01 to 2025-05-09
+// => From Operation start date to Reservation start date minus 1
+// First come 2: 2025-10-13 to 2025-10-30
+// => From Reservation end date plus 1 to Operation end date minus 1
 
 export async function createFirstComeFirstServedDateRange(season) {
   // Check if the FeatureType is "Frontcountry camping" or "Walk-in camping"
@@ -28,8 +35,8 @@ export async function createFirstComeFirstServedDateRange(season) {
       attributes: ["id", "name"],
     });
 
-    const dateTypeMap = Object.fromEntries(
-      dateTypes.map((dateType) => [dateType.id, dateType.name]),
+    const dateTypeMap = new Map(
+      dateTypes.map((dateType) => [dateType.name, dateType.id]),
     );
 
     const operationDateRange = season.dateRanges.find(
@@ -52,7 +59,7 @@ export async function createFirstComeFirstServedDateRange(season) {
 
         firstComeDateRanges.push({
           seasonId: season.id,
-          dateTypeId: dateTypeMap["First come, first served"],
+          dateTypeId: dateTypeMap.get("First come, first served"),
           startDate: firstComeStartDate,
           endDate: firstComeEndDate,
         });
@@ -61,7 +68,7 @@ export async function createFirstComeFirstServedDateRange(season) {
       // Calculate the second "First come, first served" DateRange (after Reservation)
       if (operationDateRange.endDate > reservationDateRange.endDate) {
         const firstComeStartDate = new Date(reservationDateRange.endDate);
-        
+
         firstComeStartDate.setDate(firstComeStartDate.getDate() + 1);
 
         // Subtract one day from operation end date
@@ -71,7 +78,7 @@ export async function createFirstComeFirstServedDateRange(season) {
 
         firstComeDateRanges.push({
           seasonId: season.id,
-          dateTypeId: dateTypeMap["First come, first served"],
+          dateTypeId: dateTypeMap.get("First come, first served"),
           startDate: firstComeStartDate,
           endDate: firstComeEndDate,
         });
